requirement.txt
streamlit
Pillow
torch
torchvision
pandas
numpy